Moves unlevel function alongside usage
